split Trie package
